feat(import): Allow to import from backup
Now that we can export a backup and push it to s3, we want to be able to
import it into XO, to check if the export worked.

If not, we know that we've failed somewhere.

This commit only add the command to `import` that will then be
incorporated in the archive process.