feat(WORKING_DIR): Refacto creation/rm directory
Working directory was kind of a mess.
This commit to give coherence in the way files are created and removed.
E.g, if vm metadata export fails, no need to remove the whole dir.

The whole WORKING_DIR is removed only when everything has been tested
and pushed to S3